Revamp the feature incremental update.
The previous version with a feature update stack with too
complex for something that ought to be simple. This new version
takes a much simpler approach, resuling in much fewer classes and
also lower memory and disk footprint.

The idea implemented here is that, instead of tracking when a spot
or a link is modified, we simply remove modified spots and links from
the features. This way the modified objects appear with empty values
after they have been modified (or created).

Features that want to implement incremental updates simply have to
recompute values only for objects with missing values.

In case of issues (there will be), a flag in the feature computation
service allows to force the recomputation over all objects, ignoring
their status.