Detach Hierarchy TrackScheme from the TimepointModel
There was a bug, when first linking a normal TrackScheme and a BDV view,
and afterwards linking a "Hierarchy TrackScheme" to it. Clicking a node
in a TrackScheme window would navigate to the wrong time point in the BDV
window. This bug was caused by the Hierarchy TrackSchemePanel. The Hierarchy
TrackSchemePanel works with pseudo time points, and wrote these pseudo
time points into a global TimepointModel. BDV would as a consequense show
the pseudo timepoints.

This problem is fixed there by providing the hierarchy TrackSchemePanel
with its very own "detached" TimepointModel.