Fix spot labels in the Branch TrackScheme
Now the Branch TrackScheme renders the label, that also matches the spot
time point that is used for rendering. This is achieved while keeping the
column titles and order unchanged.

It was necessary to add a getFirstLabel() to the ModelGraphProperties, this
method return the label of the first spot represented by a branch spot.
getFirstLabel is used for column titles and column sorting to be consistent
with the normal TrackSchem.

BranchSpot.getLabel() is again changed to return the label of the last spot
it represents. This is required to render the correct labels for the spots
in the BranchTrackScheme.